requirements.txt
Flask~=2.2.2
WTForms~=3.0.1
qrcode~=7.3.1
svgwrite~=1.4.3
reportlab~=3.6.11
svglib~=1.4.1
ziafont~=0.9
CairoSVG~=2.5.2
setuptools~=60.2.0
python-slugify
